requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
requests>=2.31,<3.0
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0
cryptography>=41.0,<43.0
gunicorn>=21.2,<23.0
whitenoise>=6.6,<7.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<12.0
